+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 # v2.7.45
 
+* more efficient returning in `enumerate`