+++ CHANGELOG.md
@@ -1,4 +1,4 @@
 # v2.7.45
 
-* add `measure.has_exceeded`
+* add `measure.has_exceeded` and `measure.raise_if_exceeded`