+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # v2.7.45
 
 * add `measure.has_exceeded` and `measure.raise_if_exceeded`
+* `log_exceptions` will now behave correctly when used on generators
+* `time.measure()` will now behave correctly when used on generators
+