+++ CHANGELOG.md
@@ -3,5 +3,4 @@
 * add `measure.has_exceeded` and `measure.raise_if_exceeded`
 * `log_exceptions` will now behave correctly when used on generators
 * `time.measure()` will now behave correctly when used on generators
-
-
+* fixed `count` when no start was given