+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 # v2.7.46
 
+* added `dont_check_multiple` to `choose`