+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 # v2.7.47
 
+added `random.shuffle_together`