+++ CHANGELOG.md
@@ -1,3 +1,1 @@
-# v2.7.47
-
-added `random.shuffle_together`
+# v2.7.48